+++ filter/src/config.rs
@@ -1,14 +1,14 @@
 use {
-    base64::{engine::general_purpose::STANDARD as base64_engine, Engine},
+    base64::{Engine, engine::general_purpose::STANDARD as base64_engine},
     richat_proto::geyser::{
-        subscribe_request_filter_accounts_filter::Filter as AccountsFilterDataOneof,
-        subscribe_request_filter_accounts_filter_lamports::Cmp as AccountsFilterLamports,
-        subscribe_request_filter_accounts_filter_memcmp::Data as AccountsFilterMemcmpOneof,
         CommitmentLevel as CommitmentLevelProto, SubscribeRequest,
         SubscribeRequestAccountsDataSlice, SubscribeRequestFilterAccounts,
         SubscribeRequestFilterAccountsFilter, SubscribeRequestFilterAccountsFilterLamports,
         SubscribeRequestFilterAccountsFilterMemcmp, SubscribeRequestFilterBlocks,
         SubscribeRequestFilterSlots, SubscribeRequestFilterTransactions,
+        subscribe_request_filter_accounts_filter::Filter as AccountsFilterDataOneof,
+        subscribe_request_filter_accounts_filter_lamports::Cmp as AccountsFilterLamports,
+        subscribe_request_filter_accounts_filter_memcmp::Data as AccountsFilterMemcmpOneof,
     },
     richat_shared::{
         config::{
@@ -18,8 +18,8 @@ use {
         five8::{pubkey_decode, pubkey_encode, signature_decode, signature_encode},
     },
     serde::{
-        de::{self, Deserializer},
         Deserialize, Serialize,
+        de::{self, Deserializer},
     },
     solana_commitment_config::CommitmentLevel,
     solana_sdk::{