+++ filter/src/protobuf/encode.rs
@@ -1,12 +1,12 @@
 use {
     crate::protobuf::{bytes_encode, bytes_encoded_len},
     prost::{
+        DecodeError, Message,
         bytes::{Buf, BufMut},
         encoding::{
-            self, encode_key, encode_varint, encoded_len_varint, key_len, message, DecodeContext,
-            WireType,
+            self, DecodeContext, WireType, encode_key, encode_varint, encoded_len_varint, key_len,
+            message,
         },
-        DecodeError, Message,
     },
     prost_types::Timestamp,
     richat_proto::{geyser::subscribe_update::UpdateOneof, solana::storage::confirmed_block},